+++ Voedselbank/src/voedselbank/UitgiftepuntWijzigen.java
@@ -9,6 +9,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 
 /**
  *
@@ -33,7 +34,7 @@ public class UitgiftepuntWijzigen extends javax.swing.JFrame {
         initComponents();
         this.setTitle("Uitgiftepuntgegevens wijzigen/verwijderen");
         vulVelden();
-        
+
     }
 
     public void setOpener(JFrame jf) {
@@ -55,8 +56,6 @@ public class UitgiftepuntWijzigen extends javax.swing.JFrame {
     /**
      * Maakt de combobox van uitgiftepunten aan
      */
-   
-
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -176,36 +175,45 @@ public class UitgiftepuntWijzigen extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void wijzigKnopActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_wijzigKnopActionPerformed
-        try {
-            connection = SimpleDataSourceV2.getConnection();
-            PreparedStatement ps = connection.prepareStatement("UPDATE Uitgiftepunt SET naam = ?, adres = ?, postcode = ?, plaatsnaam = ?, capaciteit = ? WHERE ID_uitgiftepunt = ?");
+        int wijzig = JOptionPane.showConfirmDialog(null, "Weet u zeker dat u de gegevens wilt wijzigen?", "Wijzigen", JOptionPane.YES_NO_OPTION);
 
-            ps.setString(1, naamVeld.getText());
-            ps.setString(2, adresVeld.getText());
-            ps.setString(3, postcodeVeld.getText());
-            ps.setString(4, plaatsnaamVeld.getText());
-            ps.setString(5, capaciteitVeld.getText());
-            ps.setInt(6, uitgiftepunt.getUitgiftepunt_ID());
+        if (wijzig == JOptionPane.YES_OPTION) {
 
-            ps.executeUpdate();
+            try {
+                connection = SimpleDataSourceV2.getConnection();
+                PreparedStatement ps = connection.prepareStatement("UPDATE Uitgiftepunt SET naam = ?, adres = ?, postcode = ?, plaatsnaam = ?, capaciteit = ? WHERE ID_uitgiftepunt = ?");
 
-        } catch (SQLException ex) {
-            ex.printStackTrace();
+                ps.setString(1, naamVeld.getText());
+                ps.setString(2, adresVeld.getText());
+                ps.setString(3, postcodeVeld.getText());
+                ps.setString(4, plaatsnaamVeld.getText());
+                ps.setString(5, capaciteitVeld.getText());
+                ps.setInt(6, uitgiftepunt.getUitgiftepunt_ID());
+
+                ps.executeUpdate();
+
+            } catch (SQLException ex) {
+                ex.printStackTrace();
+            }
         }
-
     }//GEN-LAST:event_wijzigKnopActionPerformed
 
     private void verwijderKnopActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_verwijderKnopActionPerformed
-        try {
-            connection = SimpleDataSourceV2.getConnection();
-            PreparedStatement ps = connection.prepareStatement("DELETE FROM Uitgiftepunt WHERE ID_uitgiftepunt = ?");
+        int verwijder = JOptionPane.showConfirmDialog(null, "Weet u zeker dat u alle gegevens wilt verwijderen?", "Verwijderen", JOptionPane.YES_NO_OPTION);
 
-            ps.setInt(1, uitgiftepunt.getUitgiftepunt_ID());
+        if (verwijder == JOptionPane.YES_OPTION) {
 
-            ps.executeUpdate();
+            try {
+                connection = SimpleDataSourceV2.getConnection();
+                PreparedStatement ps = connection.prepareStatement("DELETE FROM Uitgiftepunt WHERE ID_uitgiftepunt = ?");
 
-        } catch (SQLException ex) {
-            ex.printStackTrace();
+                ps.setInt(1, uitgiftepunt.getUitgiftepunt_ID());
+
+                ps.executeUpdate();
+
+            } catch (SQLException ex) {
+                ex.printStackTrace();
+            }
         }
     }//GEN-LAST:event_verwijderKnopActionPerformed
